+++ src/Reducer.ts
@@ -2,7 +2,7 @@ import { AnyAction, Reducer as ReduxReducer } from 'redux';
 import { Action, ActionType } from './types';
 
 function Reducer<TStore>(initialStore: TStore, updateActionType: ActionType) {
-	return ((store: TStore = initialStore, action: Action): TStore => {
+	return ((store: TStore = initialStore, action: Action<any>): TStore => {
 		if (action.type !== updateActionType) {
 			return store;
 		}