+++ src/index.ts
@@ -2,14 +2,7 @@ export * from './Middleware';
 export * from './controller';
 export { chainActions } from './chainActions';
 export * from './createAction';
-export {
-	Action,
-	ActionWithCallback,
-	ActionFactory,
-	isAction,
-	DecoratedWatchedController,
-	WatchedController,
-} from './types';
+export * from './types';
 export * from './Reducer';
 export * from './decorators';
 export * from './symbols';