+++ src/types/Controller.ts
@@ -1,0 +1,1 @@
+export interface Controller<State extends {} = any> {}