+++ src/types/StaticActionsCreator.ts
@@ -1,0 +1,5 @@
+import { Action } from './Action';
+
+export type StaticActionsCreator = {
+	[actionName: string]: (payload?: any) => Action;
+};